requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
decorator==4.4.2
future==0.18.2
geocoder==1.38.1
idna==2.9
newsapi-python==0.2.6
PyMsgBox==1.0.7
ratelim==0.1.6
requests==2.23.0
six==1.14.0
urllib3==1.25.8
